requirements.txt
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
Flask-SocketIO==4.2.1
PyJWT==1.7.1
psycopg2==2.8.4
praw==6.5.1
celery==4.4.2
redis==3.4.1
requests==2.23.0
requests_oauthlib==1.3.0
textblob==0.15.3
sendgrid==6.2.0
kombu==4.6.8
eventlet==0.25.1